+++ otelcollector/test/ginkgo-e2e/configprocessing/config_processing_test.go
@@ -1,7 +1,9 @@
 package configprocessing
 
 import (
+	"encoding/json"
 	"prometheus-collector/otelcollector/test/utils"
+	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -19,11 +21,11 @@ var _ = DescribeTable("The containers should be running",
 		err := utils.CheckIfAllContainersAreRunning(K8sClient, namespace, controllerLabelName, controllerLabelValue)
 		Expect(err).NotTo(HaveOccurred())
 	},
-	Entry("when checking the ama-metrics replica pod(s)", "kube-system", "rsName", "ama-metrics"),
-	Entry("when checking the ama-metrics-node", "kube-system", "dsName", "ama-metrics-node"),
-	Entry("when checking the ama-metrics-win-node pod", "kube-system", "dsName", "ama-metrics-win-node", Label(utils.WindowsLabel)),
+	Entry("when checking the ama-metrics replica pod(s)", "kube-system", "rsName", "ama-metrics", Label(utils.ConfigProcessingCommon)),
+	Entry("when checking the ama-metrics-node", "kube-system", "dsName", "ama-metrics-node", Label(utils.ConfigProcessingCommon)),
+	Entry("when checking the ama-metrics-win-node pod", "kube-system", "dsName", "ama-metrics-win-node", Label(utils.ConfigProcessingCommon)),
 	// Entry("when checking the ama-metrics-ksm pod", "kube-system", "app.kubernetes.io/name", "ama-metrics-ksm"),
-	Entry("when checking the ama-metrics-operator-targets pod", "kube-system", "rsName", "ama-metrics-operator-targets", Label(utils.OperatorLabel)),
+	Entry("when checking the ama-metrics-operator-targets pod", "kube-system", "rsName", "ama-metrics-operator-targets", Label(utils.ConfigProcessingCommon)),
 	// Entry("when checking the prometheus-node-exporter pod", "kube-system", "app", "prometheus-node-exporter", Label(utils.ArcExtensionLabel)),
 )
 
@@ -73,7 +75,7 @@ var _ = DescribeTable("otelcollector is running",
 			// "inotifywait /etc/config/settings",
 			// "inotifywait /etc/mdsd.d",
 			// "crond",
-		},
+		}, Label(utils.ConfigProcessingCommon),
 	),
 	Entry("when checking the ama-metrics-node daemonset pods", "kube-system", "dsName", "ama-metrics-node", "prometheus-collector",
 		[]string{
@@ -85,6 +87,7 @@ var _ = DescribeTable("otelcollector is running",
 			// "inotifywait /etc/mdsd.d",
 			// "crond",
 		},
+		Label(utils.ConfigProcessingCommon),
 	),
 )
 
@@ -106,7 +109,7 @@ var _ = DescribeTable("otelcollector is running",
 			// "MonAgentManager",
 			// "MonAgentCore",
 		},
-		Label(utils.WindowsLabel),
+		Label(utils.ConfigProcessingCommon),
 		FlakeAttempts(3),
 	),
 )
@@ -122,13 +125,13 @@ var _ = DescribeTable("The container logs should not contain errors",
 		err := utils.CheckContainerLogsForErrors(K8sClient, namespace, controllerLabelName, controllerLabelValue)
 		Expect(err).NotTo(HaveOccurred())
 	},
-	Entry("when checking the ama-metrics replica pods", "kube-system", "rsName", "ama-metrics"),
-	Entry("when checking the ama-metrics-node", "kube-system", "dsName", "ama-metrics-node"),
-	Entry("when checking the ama-metrics replica pods", "kube-system", "rsName", "ama-metrics", Label(utils.ARM64Label)),
-	Entry("when checking the ama-metrics-node", "kube-system", "dsName", "ama-metrics-node", Label(utils.ARM64Label)),
-	Entry("when checking the ama-metrics-win-node", "kube-system", "dsName", "ama-metrics-win-node", Label(utils.WindowsLabel)),
+	Entry("when checking the ama-metrics replica pods", "kube-system", "rsName", "ama-metrics", Label(utils.ConfigProcessingCommon)),
+	Entry("when checking the ama-metrics-node", "kube-system", "dsName", "ama-metrics-node", Label(utils.ConfigProcessingCommon)),
+	Entry("when checking the ama-metrics replica pods", "kube-system", "rsName", "ama-metrics", Label(utils.ConfigProcessingCommon)),
+	Entry("when checking the ama-metrics-node", "kube-system", "dsName", "ama-metrics-node", Label(utils.ConfigProcessingCommon)),
+	Entry("when checking the ama-metrics-win-node", "kube-system", "dsName", "ama-metrics-win-node", Label(utils.ConfigProcessingCommon)),
 	// Entry("when checking the ama-metrics-ksm pod", "kube-system", "app.kubernetes.io/name", "ama-metrics-ksm"),
-	Entry("when checking the ama-metrics-operator-targets pod", "kube-system", "rsName", "ama-metrics-operator-targets", Label(utils.OperatorLabel)),
+	Entry("when checking the ama-metrics-operator-targets pod", "kube-system", "rsName", "ama-metrics-operator-targets", Label(utils.ConfigProcessingCommon)),
 	// Entry("when checking the prometheus-node-exporter pod", "kube-system", "app", "prometheus-node-exporter", Label(utils.ArcExtensionLabel)),
 )
 
@@ -136,38 +139,38 @@ var _ = DescribeTable("The container logs should not contain errors",
  * Test that the Prometheus UI /config API endpoint returns a Prometheus config that can be unmarshaled.
  */
 // All targets disabled
-// var _ = DescribeTable("The Prometheus UI API should return empty config",
-// 	func(namespace string, controllerLabelName string, controllerLabelValue string, containerName string, isLinux bool) {
-// 		time.Sleep(120 * time.Second)
-// 		var apiResponse utils.APIResponse
-// 		err := utils.QueryPromUIFromPod(K8sClient, Cfg, namespace, controllerLabelName, controllerLabelValue, containerName, "/api/v1/status/config", isLinux, &apiResponse)
-// 		Expect(err).NotTo(HaveOccurred())
-// 		Expect(apiResponse.Data).NotTo(BeNil())
+var _ = DescribeTable("The Prometheus UI API should return empty config",
+	func(namespace string, controllerLabelName string, controllerLabelValue string, containerName string, isLinux bool) {
+		time.Sleep(120 * time.Second)
+		var apiResponse utils.APIResponse
+		err := utils.QueryPromUIFromPod(K8sClient, Cfg, namespace, controllerLabelName, controllerLabelValue, containerName, "/api/v1/status/config", isLinux, &apiResponse)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(apiResponse.Data).NotTo(BeNil())
 
-// 		// // var prometheusConfigResult v1.ConfigResult
-// 		// var prometheusConfigResult map[string]interface{}
-// 		// json.Unmarshal([]byte(apiResponse.Data), &prometheusConfigResult)
-// 		// Expect(prometheusConfigResult).NotTo(BeNil())
-// 		// //Expect(prometheusConfigResult.YAML).NotTo(BeEmpty())
-// 		// scrapeConfigs := prometheusConfigResult["scrape_configs"]
-// 		// //prometheusConfig, err := config.Load(prometheusConfigResult.YAML, true, nil)
-// 		// Expect(err).NotTo(HaveOccurred())
-// 		// Expect(scrapeConfigs).NotTo(BeNil())
+		// // var prometheusConfigResult v1.ConfigResult
+		// var prometheusConfigResult map[string]interface{}
+		// json.Unmarshal([]byte(apiResponse.Data), &prometheusConfigResult)
+		// Expect(prometheusConfigResult).NotTo(BeNil())
+		// //Expect(prometheusConfigResult.YAML).NotTo(BeEmpty())
+		// scrapeConfigs := prometheusConfigResult["scrape_configs"]
+		// //prometheusConfig, err := config.Load(prometheusConfigResult.YAML, true, nil)
+		// Expect(err).NotTo(HaveOccurred())
+		// Expect(scrapeConfigs).NotTo(BeNil())
 
-// 		var prometheusConfigResult v1.ConfigResult
-// 		json.Unmarshal([]byte(apiResponse.Data), &prometheusConfigResult)
-// 		Expect(prometheusConfigResult).NotTo(BeNil())
-// 		Expect(prometheusConfigResult.YAML).NotTo(BeEmpty())
+		var prometheusConfigResult v1.ConfigResult
+		json.Unmarshal([]byte(apiResponse.Data), &prometheusConfigResult)
+		Expect(prometheusConfigResult).NotTo(BeNil())
+		Expect(prometheusConfigResult.YAML).NotTo(BeEmpty())
 
-// 		prometheusConfig, err := config.Load(prometheusConfigResult.YAML, true, nil)
-// 		Expect(err).NotTo(HaveOccurred())
-// 		Expect(prometheusConfig).NotTo(BeNil())
-// 		Expect(prometheusConfig.ScrapeConfigs).NotTo(BeNil())
-// 		Expect(len(prometheusConfig.ScrapeConfigs)).To(BeNumerically("==", 1))
-// 		Expect(prometheusConfig.ScrapeConfigs[0].JobName).To(Equal("empty_job"))
+		prometheusConfig, err := config.Load(prometheusConfigResult.YAML, true, nil)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(prometheusConfig).NotTo(BeNil())
+		Expect(prometheusConfig.ScrapeConfigs).NotTo(BeNil())
+		Expect(len(prometheusConfig.ScrapeConfigs)).To(BeNumerically("==", 1))
+		Expect(prometheusConfig.ScrapeConfigs[0].JobName).To(Equal("empty_job"))
 
-// 	},
-// 	Entry("when called inside ama-metrics replica pod", "kube-system", "rsName", "ama-metrics", "prometheus-collector", true),
-// 	Entry("when called inside the ama-metrics-node pod", "kube-system", "dsName", "ama-metrics-node", "prometheus-collector", true),
-// 	Entry("when checking the ama-metrics-win-node", "kube-system", "dsName", "ama-metrics-win-node", "prometheus-collector", false, Label(utils.WindowsLabel)),
-// )
+	},
+	Entry("when called inside ama-metrics replica pod", "kube-system", "rsName", "ama-metrics", "prometheus-collector", true, Label(utils.ConfigProcessingAllTargetsDisabled)),
+	Entry("when called inside the ama-metrics-node pod", "kube-system", "dsName", "ama-metrics-node", "prometheus-collector", true, Label(utils.ConfigProcessingAllTargetsDisabled)),
+	Entry("when checking the ama-metrics-win-node", "kube-system", "dsName", "ama-metrics-win-node", "prometheus-collector", false, Label(utils.ConfigProcessingAllTargetsDisabled)),
+)
